feat: search for courses filtering modalities
Courses can be showed based on not just its own attributes but modalities associated with them. Ex: show all courses with degree as Bachelor which modalities are On-campus. Also, it's important to mention that the filtering on query params separated by comma is possible due to 'query-string' lib, which can parses and identify arrays with this pattern.