fix: ra and semester validation on student
It's possible now for a student RA be a string containing numbers with 0 as first digit. Also, the semester was always being invalid on index method because every data obtained from query params is a string, so on the field's validation it was needed to ensure that semester was in fact a number.