refac: remove useless default value on destructuring
The default values of variables on destructuring assignments are useless. No error is thrown iterating on an array with undefined or null as one of its values; just is thrown when trying to use a for loop with null or undefined, which are not iterable (that's why i kept 'let { modality = "" }' on course update.