feat: specify predefined fks on M:N associations
to enforce that sequelize, in the same way as the other models, know which is the foreign keys on each association belongsToMany that establish the M:N relation between course and modality / user and permission. This way all the models names could be changed into PascalCase again and no unexpected issues would happen with sequelize considering foreign key names as PascalCase too. (probably it won't because this can change all included data field names to PascalCase and the name prop of all models too, and this matters).